+++ src/core/base.rs
@@ -10,7 +10,7 @@ pub trait Obs: Clone + Debug {
 
     /// Replace elements of observation where `is_done[i] == 1.0`.
     /// This method assumes that `is_done.len() == n_procs`.
-    fn merge(self, obs_reset: Self, is_done: &[f32]) -> Self;
+    fn merge(self, obs_reset: Self, is_done: &[i8]) -> Self;
 }
 
 /// Represents an action of the environment.
@@ -25,12 +25,12 @@ pub struct Step<E: Env> {
     pub act: E::Act,
     pub obs: E::Obs,
     pub reward: Vec<f32>,
-    pub is_done: Vec<f32>,
+    pub is_done: Vec<i8>,
     pub info: E::Info,
 }
 
 impl<E: Env> Step<E> {
-    pub fn new(obs: E::Obs, act: E::Act, reward: Vec<f32>, is_done: Vec<f32>, info: E::Info) -> Self {
+    pub fn new(obs: E::Obs, act: E::Act, reward: Vec<f32>, is_done: Vec<i8>, info: E::Info) -> Self {
         Step {
             act,
             obs,
@@ -51,7 +51,7 @@ pub trait Env {
 
     /// Reset the i-th environment if `is_done[i]==1.0`.
     /// Thei-th return value should be ignored if `is_done[i]==0.0`.
-    fn reset(&mut self, is_done: Option<&Vec<f32>>) -> Result<Self::Obs, Box<dyn Error>>;
+    fn reset(&mut self, is_done: Option<&Vec<i8>>) -> Result<Self::Obs, Box<dyn Error>>;
 }
 
 /// Represents a policy. on an environment. It is based on a mapping from an observation
